feat: add store & slice
