+++ src/components/all-providers.tsx
@@ -1,6 +1,12 @@
+import { store } from '@/libs/store';
 import { router } from '@routes/router';
+import { Provider } from 'react-redux';
 import { RouterProvider } from 'react-router-dom';
 
 export default function AllProviders() {
-  return <RouterProvider router={router} />;
+  return (
+    <Provider store={store}>
+      <RouterProvider router={router} />;
+    </Provider>
+  );
 }
