+++ src/components/uncontrolled.tsx
@@ -3,95 +3,27 @@ import * as Yup from 'yup';
 
 import styles from '@styles/form.module.css';
 import global from '@styles/global.module.css';
-
-export interface FormValues {
-  name: string;
-  age: string;
-  email: string;
-  password: string;
-  repeat: string;
-  country: string;
-  gender: string;
-  upload: File;
-  accept: boolean;
-}
-
-interface FormErrors {
-  name?: string;
-  age?: string;
-  email?: string;
-  password?: string;
-  repeat?: string;
-  country?: string;
-  gender?: string;
-  upload?: string;
-  accept?: boolean;
-}
-
-const d = [
-  'Type name',
-  'Type age',
-  'Type email',
-  'Type password',
-  'Repeat password',
-  'Start typing county',
-];
-
-const fileTypes = ['image/jpeg', 'image/png'];
-
-const schema = Yup.object().shape({
-  name: Yup.string()
-    .matches(/^[A-Z]/, 'Name must start with an uppercase letter')
-    .required()
-    .label('Name'),
-  age: Yup.number()
-    .transform((value: string) => (Number.isNaN(value) ? null : value))
-    .positive()
-    .integer()
-    .required('Age must be a number')
-    .label('Age'),
-  email: Yup.string().email().required().label('Email'),
-  password: Yup.string()
-    .min(6, 'Password must be at least 6 characters long')
-    .matches(/[A-Z]/, 'Password contain at least one uppercase letter')
-    .matches(/[a-z]/, 'Password contain at least one lowercase letter')
-    .matches(/[0-9]/, 'Password must contain at least one number')
-    .matches(
-      /[!@#$%^&*(),.?":{}|<>]/,
-      'Password contain at least one special character'
-    )
-    .required()
-    .label('Password'),
-  repeat: Yup.string()
-    .oneOf([Yup.ref('password')], 'Password must match')
-    .required()
-    .label('Password'),
-  country: Yup.string().required().label('Country'),
-  gender: Yup.string().required().label('Gender'),
-  upload: Yup.mixed<File>()
-    .test('fileSize', 'File size must be less than 1MB', (value) => {
-      return value && value.size <= 1024 * 1024;
-    })
-    .test('fileType', 'File extension must be .jpg, .jpeg, .png', (value) => {
-      return value && fileTypes.includes(value.type);
-    })
-    .required()
-    .label('Image File'),
-  accept: Yup.string()
-    .required('Please accept Terms and Conditions')
-    .label('Terms'),
-});
+import { placeholderDescription as d } from '@/libs/const';
+import { FormErrors, FormValue } from '@/libs/interfaces';
+import { schemaValidation } from '@/libs/schemaValidation';
+import { useAppDispatch } from '@/libs/store-hooks';
+import { addFormValue } from '@/libs/store';
 
 export default function Uncontrolled() {
   const [errors, setErrors] = useState<FormErrors>({});
+  const dispatch = useAppDispatch();
 
   async function onSubmit(event: FormEvent<HTMLFormElement>) {
+    setErrors({});
     const formData = new FormData(event.currentTarget);
     const formObject = Object.fromEntries(formData.entries());
 
     try {
-      await schema.validate(formObject, { abortEarly: false });
-      setErrors({});
+      await schemaValidation.validate(formObject, { abortEarly: false });
+      const upload = URL.createObjectURL(formObject.upload as File);
+      dispatch(
+        addFormValue({ ...(formObject as unknown as FormValue), upload })
+      );
     } catch (validationErrors) {
       const newErrors: Record<string, string> = {};
