+++ src/libs/const.ts
@@ -1,0 +1,8 @@
+export const placeholderDescription = [
+  'Type name',
+  'Type age',
+  'Type email',
+  'Type password',
+  'Repeat password',
+  'Start typing county',
+];
