+++ src/libs/interfaces.ts
@@ -1,0 +1,23 @@
+export interface FormValue {
+  name: string;
+  age: string;
+  email: string;
+  password: string;
+  repeat: string;
+  country: string;
+  gender: string;
+  upload?: string;
+  accept: string;
+}
+
+export interface FormErrors {
+  name?: string;
+  age?: string;
+  email?: string;
+  password?: string;
+  repeat?: string;
+  country?: string;
+  gender?: string;
+  upload?: string;
+  accept?: string;
+}
