+++ vite.config.ts
@@ -12,6 +12,7 @@ export default defineConfig({
       '@components': path.resolve(__dirname, './src/components'),
       '@styles': path.resolve(__dirname, './src/styles'),
       '@routes': path.resolve(__dirname, './src/routes'),
+      '@libs': path.resolve(__dirname, './src/libs'),
     },
   },
   plugins: [react(), magicalSvg({ target: 'react-jsx' })],
